+++ data_pipelines_cli/cli_commands/test.py
@@ -1,5 +1,6 @@
 import click
 
+from ..config_generation import generate_profiles_yml
 from ..dbt_utils import run_dbt_command
 
 
@@ -10,7 +11,9 @@ def test(env: str) -> None:
     :param env: Name of the environment
     :type env: str
     """
-    run_dbt_command(("test",), env, None)
+    profiles_path = generate_profiles_yml(env)
+    run_dbt_command(("deps",), env, profiles_path)
+    run_dbt_command(("test",), env, profiles_path)
 
 
 @click.command(name="test", help="Run tests of the project on the local machine")