+++ data_pipelines_cli/dbt_utils.py
@@ -1,15 +1,12 @@
 import pathlib
 import sys
-from typing import Any, Dict, Optional, Tuple
+from typing import Any, Dict, Tuple
 
 import yaml
 
 from .cli_constants import BUILD_DIR, get_dbt_profiles_env_name
 from .cli_utils import echo_subinfo, subprocess_run
-from .config_generation import (
-    generate_profiles_yml,
-    read_dictionary_from_config_directory,
-)
+from .config_generation import read_dictionary_from_config_directory
 from .data_structures import DataPipelinesConfig, read_config
 
 
@@ -23,7 +20,7 @@ def _read_dbt_vars_from_configs(dbt_env_config: Dict[str, Any]) -> str:
 
 
 def run_dbt_command(
-    command: Tuple[str, ...], env: str, profiles_path: Optional[pathlib.Path]
+    command: Tuple[str, ...], env: str, profiles_path: pathlib.Path
 ) -> None:
     """
     Runs dbt subprocess in a context of specified *env*
@@ -33,12 +30,11 @@ def run_dbt_command(
     :param env: Name of the environment
     :type env: str
     :param profiles_path: Path to the directory containing `profiles.yml` file
-    :type profiles_path: Optional[pathlib.Path]
+    :type profiles_path: pathlib.Path
     """
     command_str = " ".join(list(command))
     echo_subinfo(f"dbt {command_str}")
 
-    profiles_path = profiles_path or generate_profiles_yml(env)
     dbt_env_config = read_dictionary_from_config_directory(
         BUILD_DIR.joinpath("dag"), env, "dbt.yml"
     )