+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - `--env` option has a default value: `local`,
   - `--datahub` is changed to `--datahub-gms-uri`, `--repository` is changed to `--docker-repository-uri`.
 - `dp deploy`'s `--docker-push` is not a flag anymore and requires a Docker repository URI parameter; `--repository` got removed then.
+- `dp run` and `dp test` run `dbt deps` before actual **dbt** command.
 
 ## [0.6.0] - 2021-12-16